requirements.txt
Django
pytest
pytest-pythonpath
pytest-django
djangorestframework
djangorestframework-simplejwt
Pillow
PyJWT
requests
